# services/authentication/delivery/http/handlers/refresh_token_handler.go
package handlers

import (
	httpLib "libraries/http"
	"libraries/http/json"
	"net/http"
	"services/authentication/delivery/http/dto"
	"services/authentication/service"
)

type RefreshTokenHandler struct {
	service *service.AuthenticationService
}

func NewRefreshTokenHandler(service *service.AuthenticationService) *RefreshTokenHandler {
	return &RefreshTokenHandler{service: service}
}

// Handle godoc
// @Summary Refresh access token
// @Description Exchanges refresh token for new access token.
// @Tags Authentication
// @Accept json
// @Produce json
// @Param request body dto.RefreshTokenRequest true "Refresh token request"
// @Success 200 {object} dto.TokenResponse
// @Failure 400 {object} dto.ErrorResponse
// @Failure 500 {object} dto.ErrorResponse
// @Router /v1/authentication/refresh [post]
func (h *RefreshTokenHandler) Handle(w http.ResponseWriter, r *http.Request) {
	var req dto.RefreshTokenRequest
	if err := json.Read(r, &req); err != nil {
		httpLib.HandleErrorWithStatus(w, http.StatusBadRequest, err.Error())
		return
	}

	token, err := h.service.RefreshToken(r.Context(), req.RefreshToken)
	if err != nil {
		httpLib.HandleError(w, err)
		return
	}

	response := dto.TokenResponse{
		AccessToken:  token.AccessToken,
		RefreshToken: token.RefreshToken,
		IDToken:      token.IDToken,
		ExpiresIn:    token.ExpiresIn,
		TokenType:    token.TokenType,
	}

	json.Write(w, http.StatusOK, response)
}